refactor: split advisor output helpers
